### 2021-01-30 #2998 @whedon
[PRE REVIEW]: Generating fragment density plots in R/Bioconductor with VplotR
**Submitting author:** @js2264 (<a href="http://orcid.org/0000-0002-4295-0624">Jacques Serizay</a>)
**Repository:** <a href="https://github.com/js2264/VplotR/" target ="_blank">https://github.com/js2264/VplotR/</a>
**Version:** v1.1.0
