### 2021-01-30 #2998 @js2264
> @js2264 - I see a mention of a license in your readme, but I don't see a license file in the repository

Thanks for spotting it. Not sure why it was absent, but should be fixed now (both in JOSS and master branch)!
