### 2021-02-04 #3009 @csoneson
👋🏼 @js2264, @henrykironde, @fgeier - this is the review thread for the paper. All of our communications will happen here from now on.

Both reviewers have checklists at the top of this thread. As you go over the submission, please check any items that you feel have been satisfied. There are also links to the JOSS reviewer guidelines.
