### 2021-02-06 js2264/VplotR#4 @henrykironde
Revisions for Joss submission
### Functionality

I tried the installation on MAC os  but failed with the following errors.
```R
* installing *source* package ‘VplotR’ ...
** using staged installation
** R
** data
** inst
** byte-compile and prepare package for lazy loading
Error: (converted from warning) package ‘GenomicRanges’ was built under R version 4.0.3
Execution halted
ERROR: lazy loading failed for package ‘VplotR’
* removing ‘/Library/Frameworks/R.framework/Versions/4.0/Resources/library/VplotR’
Error: Failed to install 'VplotR' from GitHub:
  (converted from warning) installation of package ‘/var/folders/d5/vm7jfw2x7q550xrltd8ss5440000gn/T//RtmpFMzoYK/file23021e4afa6d/VplotR_1.1.0.tar.gz’ had non-zero exit status
```

And on another Mac os, the software installed but failed to load with the following error.
```R
* installing *source* package ‘VplotR’ ...
** using staged installation
** R
** data
** inst
** byte-compile and prepare package for lazy loading
** help
*** installing help indices
** building package indices
** installing vignettes
** testing if installed package can be loaded from temporary location
** testing if installed package can be loaded from final location
** testing if installed package keeps a record of temporary installation path
* DONE (VplotR)
> library(VplotR)
Loading required package: GenomicRanges
Loading required package: stats4
Loading required package: BiocGenerics
Loading required package: parallel

Attaching package: ‘BiocGenerics’

The following objects are masked from ‘package:parallel’:

    clusterApply, clusterApplyLB, clusterCall, clusterEvalQ,
    clusterExport, clusterMap, parApply, parCapply, parLapply,
    parLapplyLB, parRapply, parSapply, parSapplyLB

The following objects are masked from ‘package:stats’:

    IQR, mad, sd, var, xtabs

The following objects are masked from ‘package:base’:

    anyDuplicated, append, as.data.frame, basename, cbind, colnames,
    dirname, do.call, duplicated, eval, evalq, Filter, Find, get, grep,
    grepl, intersect, is.unsorted, lapply, Map, mapply, match, mget,
    order, paste, pmax, pmax.int, pmin, pmin.int, Position, rank,
    rbind, Reduce, rownames, sapply, setdiff, sort, table, tapply,
    union, unique, unsplit, which.max, which.min

Loading required package: S4Vectors

Attaching package: ‘S4Vectors’

The following object is masked from ‘package:base’:

    expand.grid

Loading required package: IRanges
Loading required package: GenomeInfoDb
Loading required package: ggplot2
> getFragmentsDistribution()
Error in get0(oNam, envir = ns) :
  lazy-load database '/Library/Frameworks/R.framework/Versions/4.0/Resources/library/magrittr/R/magrittr.rdb' is corrupt
In addition: Warning message:
In get0(oNam, envir = ns) : internal error -3 in R_decompress1
```

The Documentation shows another approach of installing the software, could we have this also mention in the README.md
if(!requireNamespace("BiocManager", quietly = TRUE))
install.packages("BiocManager")
BiocManager::install("VplotR")
library("VplotR")

I would prefer more explanation of the functions on the README page. This is the landing page, we should have a simple get started with the tools. For example we can copy the functions mentioned below from the Docs to the read me

```
Importing sequencing datasets
Provided datasets
Single Vplot
```

Also an explanation of where the datasets come from and how we can view the list of the datasets would help users

### Documentation

Add the link of the Documentation on to the README page. I had to look for the doc page https://jserizay.com/VplotR/articles/VplotR.html

I recommend adding a code of conduct page.
Example: https://github.com/ropensci/rdataretriever/blob/master/CODE_OF_CONDUCT.md
