### 2021-02-06 #3009 @henrykironde
@js2264, I have made some recommendations as issues on the repo. https://github.com/js2264/VplotR/issues/4. Please let me know if there is something not clear as I continue to review.
