+++ README.md
@@ -26,9 +26,23 @@ makes full use of its potential. As such, it is easy to generate V-plots
 in batches and combine them with other plots to make 
 publication-ready figures.  
 
+## System requirements 
+
+VplotR has been tested on Mac OS (`>= 10.13`), Ubuntu (`>= 18.04`) and Windows Server (`2012`).  
+VplotR requires `R >= 4.0` and is available from Bioconductor (release `> 3.12`). 
+
 ## Installation
 
-VplotR can be installed from Github as follows:
+VplotR and all its dependencies can be installed from Bioconductor:
+
+```r
+if (!requireNamespace("BiocManager", quietly = TRUE))
+    install.packages("BiocManager")
+BiocManager::install("VplotR")
+library("VplotR")
+```
+
+VplotR developpment version can be installed from Github as follows:
 
 ```r
 install.packages("devtools")
@@ -48,14 +62,105 @@ The main user-level functions of VplotR are `getFragmentsDistribution()`,
 * `plotProfile()` is used to plot the distribution of paired-end fragments 
   at a single locus of interest.
 
-See the package vignettes for full description of each function 
-and examples of use cases.
+See the [package vignette](https://jserizay.com/VplotR/articles/VplotR.html) 
+for an in-depth introduction to the package functionalities and 
+the [reference](https://jserizay.com/VplotR/reference/index.html) 
+for a full description of each function, a list of the available data 
+provided in the package and examples of use cases.
+
+## Get started with VplotR 
+
+### Plotting a V-plot
+
+V-plots can be generated using the `plotVmat()` function: 
+
+```r
+data(MNase_sacCer3_Henikoff2011)
+data(ABF1_sacCer3)
+p <- plotVmat(
+    x = MNase_sacCer3_Henikoff2011, 
+    granges = ABF1_sacCer3
+)
+```
+
+### Importing sequencing fragments are `GRanges`
+
+Paired-end `.bam` files are imported using the `importPEBamFiles()` 
+function as follows:
+
+```r
+library(VplotR)
+bamfile <- system.file("extdata", "ex1.bam", package = "Rsamtools")
+fragments <- importPEBamFiles(
+    bamfile,
+    shift_ATAC_fragments = TRUE
+)
+```
+
+### Plotting fragment size distribution 
+
+The distribution of fragment sizes can be computed with 
+the `getFragmentsDistribution()` function: 
+
+```r
+library(VplotR)
+data(MNase_sacCer3_Henikoff2011)
+data(ABF1_sacCer3)
+df <- getFragmentsDistribution(
+    MNase_sacCer3_Henikoff2011,
+    ABF1_sacCer3
+)
+ggplot(df, aes(x = x, y = y)) + 
+    geom_line() + 
+    theme_ggplot2()
+```
+
+### Datasets provided in VplotR
+
+Several dataset samples are provided as `GRanges` 
+object in this package:
+
+- A set of tissue-specific ATAC-seq experiments in
+young adult _C. elegans_ (Serizay et al., 2020):
+
+```r
+library(VplotR)
+data(ATAC_ce11_Serizay2020)
+```
+
+- An MNase-seq experiment in yeast (Henikoff et al., 2011) 
+and associated ABF1 binding sites:
+
+```r
+library(VplotR)
+data(MNase_sacCer3_Henikoff2011)
+```
+
+- A `.bam` file sample containing 100,000 reads mapped to _C. elegans_:
+
+```r
+library(VplotR)
+data(bam_test)
+```
+
+For a full list of the data available in this package, please go to the 
+[reference page](https://jserizay.com/VplotR/reference/index.html#section-data). 
 
 ## Contributions
+
 Code contributions, bug reports, fixes and feature requests are most welcome.
[…]
