+++ .travis.yml
@@ -1,24 +1,0 @@
-language: r
-cache: packages
-warnings_are_errors: false
-notifications:
-  email:
-    on_success: change
-    on_failure: change
-matrix:
-  include:
-    - os: linux
-      dist: trusty
-      sudo: true
-      r: bioc-devel
-      after_success: 
-        - Rscript -e 'covr::codecov()'
-        - Rscript -e 'pkgdown::build_site()'
-deploy:
-  provider: pages
-  skip-cleanup: true
-  github-token: $GITHUB_PAT
-  keep-history: true
-  local-dir: docs
-  on:
-    branch: master
