+++ pkgdown.yml
@@ -1,6 +1,7 @@
-pandoc: '2.2'
-pkgdown: 1.5.1
+pandoc: 2.7.3
+pkgdown: 1.6.1
 pkgdown_sha: ~
-articles: []
-last_built: 2020-08-14T18:37Z
+articles:
+  VplotR: VplotR.html
+last_built: 2021-02-06T19:32Z
 
