+++ .github/workflows/draft-pdf.yml
@@ -1,7 +1,4 @@
-on:
-  push:
-    branches:
-      - JOSS
+on: [push]
 
 jobs:
   paper:
