### 2021-02-06 js2264/VplotR#4 @js2264
I have addressed these issues in my last commits.

Regarding the installation, the package is automatically tested (both its installation and the multiple tests implemented) by nightly Bioconductor builds (reports here: http://bioconductor.org/checkResults/release/bioc-LATEST/VplotR/), so if the package is installed from Bioconductor (`BiocManager::install("VplotR")`), the install should be straightforward and all functions should work (builds are tested on Mac 10.14.6, Ubuntu 18.04 and 20.04 and Windows Server 2012).

Please let me know if after installing it from Bioconductor, you are still unable to follow the vignette.
