### 2021-02-06 e2ae2cd
add plots to README
