+++ README.md
@@ -83,6 +83,8 @@ p <- plotVmat(
 )
 ```
 
+![vplot](inst/figures/vplot.png)
+
 ### Importing sequencing fragments are `GRanges`
 
 Paired-end `.bam` files are imported using the `importPEBamFiles()` 
@@ -95,6 +97,26 @@ fragments <- importPEBamFiles(
     bamfile,
     shift_ATAC_fragments = TRUE
 )
+fragments
+```
+
+```r
+##   GRanges object with 1572 ranges and 0 metadata columns:
+##            seqnames    ranges strand
+##               <Rle> <IRanges>  <Rle>
+##        [1]     seq1    41-215      +
+##        [2]     seq1    54-255      +
+##        [3]     seq1    56-258      +
+##        [4]     seq1    65-255      +
+##        [5]     seq1    65-265      +
+##        ...      ...       ...    ...
+##     [1568]     seq2 1326-1542      -
+##     [1569]     seq2 1336-1544      -
+##     [1570]     seq2 1358-1550      -
+##     [1571]     seq2 1380-1557      -
+##     [1572]     seq2 1353-1562      -
+##     -------
+##     seqinfo: 2 sequences from an unspecified genome; no seqlengths
 ```
 
 ### Plotting fragment size distribution 
@@ -112,8 +134,11 @@ df <- getFragmentsDistribution(
 )
 ggplot(df, aes(x = x, y = y)) + 
     geom_line() + 
-    theme_ggplot2()
+    theme_ggplot2() + 
+    labs(x = "Fragment size", y = "# of fragments")
 ```
+
+![fragment_size_distribution](inst/figures/fragment_size_distribution.png)
 
 ### Datasets provided in VplotR
 
@@ -126,6 +151,45 @@ young adult _C. elegans_ (Serizay et al., 2020):
 ```r
 library(VplotR)
 data(ATAC_ce11_Serizay2020)
+ATAC_ce11_Serizay2020
+```
+
+```r
+##   $Germline
+##   GRanges object with 462371 ranges and 0 metadata columns:
+##              seqnames            ranges strand
+##                 <Rle>         <IRanges>  <Rle>
+##          [1]     chrI           426-514      +
+##          [2]     chrI         3588-3854      +
+##          [3]     chrI         3640-3798      +
+##          [4]     chrI         3650-3694      +
+##          [5]     chrI         3732-3863      +
+##          ...      ...               ...    ...
+##     [462367]     chrX 17712277-17712469      -
+##     [462368]     chrX 17712279-17712342      -
+##     [462369]     chrX 17712282-17712565      -
+##     [462370]     chrX 17712285-17712384      -
+##     [462371]     chrX 17712287-17712576      -
+##     -------
+##     seqinfo: 7 sequences from an unspecified genome; no seqlengths
+##   
+##   $Neurons
+##   GRanges object with 367935 ranges and 0 metadata columns:
+##              seqnames            ranges strand
+##                 <Rle>         <IRanges>  <Rle>
+##          [1]     chrI         4011-4241      +
+##          [2]     chrI         7397-7614      +
+##          [3]     chrI       11279-11502      +
+##          [4]     chrI       12744-12819      +
+##          [5]     chrI       14381-14433      +
+##          ...      ...               ...    ...
+##     [367931]     chrX 17687948-17687982      -
+##     [367932]     chrX 17699614-17699853      -
+##     [367933]     chrX 17706798-17706923      -
+##     [367934]     chrX 17708264-17708347      -
+##     [367935]     chrX 17709920-17710007      -
+##     -------
+##     seqinfo: 7 sequences from an unspecified genome; no seqlengths
 ```
 
 - An MNase-seq experiment in yeast (Henikoff et al., 2011) 
[…]
