### 2021-02-07 #3009 @csoneson
> @csoneson I have merged JOSS branch to master branch. Should I do anything for whedon to know about this change?

No, I think this should be all good. And whedon seems to find the paper in the master branch
