### 2021-02-07 d2954b0
improve README and link to vignette on top
