+++ pkgdown.yml
@@ -3,5 +3,5 @@ pkgdown: 1.6.1
 pkgdown_sha: ~
 articles:
   VplotR: VplotR.html
-last_built: 2021-02-06T22:26Z
+last_built: 2021-02-07T09:52Z
 
