### 2021-02-10 #3009 @henrykironde
Thanks for the update. I checked the issues and they look good. I am trying to evaluate the paper at the moment. I know the total number of words exceed the maximum (1000) by ~120 words. I will have to get some advice on that. For now, I will go through the writing and get back to the checkboxes.
