### 2021-02-22 #3009 @csoneson
@js2264 - you could perhaps get a DOI for your grant application by submitting your paper to a [preprint server](https://joss.readthedocs.io/en/latest/submitting.html#preprint-policy). Of course it is possible that you will have a JOSS DOI by the end of the week, but I don't think one should assume such a rapid turnaround time (note that, in particular, JOSS is still operating in [reduced service mode](https://blog.joss.theoj.org/2020/05/reopening-joss) because of the challenges brought by the pandemic).
