### 2021-02-24 #3009 @js2264
:wave: @fgeier I checked with the grant selection committee and they told me that although I can list preprints in my application, accepted papers would be a major plus. And as you said to tell you if I needed it quicker, well, I thought I would let you know. But again, if you cannot make it by the end of the week then I would definitely understand! The reviewing process should not be rushed because of a deadline, so take the time you need. Thanks again :smile:
