### 2021-02-24 #3009 @fgeier
@js2264 I added a comment on `plotProfile` as an issue in your repo. Please address it. Otherwise I am fine with publication. Its a certainly a very helpful package, not only for ATAC data. I must say however that some performance improvement of `importPEBamFiles` would really benefit the analysis. I tried the analysis of a 10GB ATAC dataset on 22k TSS regions and it essentially didn't complete within 4 hours (after which I stopped the analysis). In case of a single bam input you could maybe parallelize the import over seqlevels?
