### 2021-02-25 a190296
improved description for plotProfile
