+++ paper/paper.bib
@@ -27,7 +27,7 @@
 }
 
 @article{Serizay2020Oct,
-	author = {Serizay, Jacques and Dong, Yan and J{\ifmmode\ddot{a}\else\"{a}\fi}nes, J{\ifmmode\ddot{u}\else\"{u}\fi}rgen and Chesney, Michael and Cerrato, Chiara and Ahringer, Julie},
+	author = {Serizay, Jacques and Dong, Yan and Jänes, Jürgen and Chesney, Michael and Cerrato, Chiara and Ahringer, Julie},
 	title = {{Distinctive regulatory architectures of germline-active and somatic genes in C. elegans}},
 	journal = {Genome Research},
 	year = {2020},
