+++ man/plotProfile.Rd
@@ -43,7 +43,13 @@ this locus will be in red.}
 A ggplot
 }
 \description{
-A function to generate a Vplot along chromosome coordinates
+The paired-end fragments overlapping a locus of interest 
+(e.g., binding sites, provided in the `loci` argument) are shown in red 
+while the remaining fragments mapping to the genomic window are 
+displayed in black. 
+Marginal curves are also plotted on the side of the distribution plot. 
+They highlight the smoothed distribution of the position of paired-end 
+fragment midpoints (top) or of the paired-end fragment length (right)
 }
 \examples{
 data(bam_test)
