### 2021-02-25 #3009 @js2264
@fgeier Thank you so much for taking the time to go through all the review process! I have now improved the doc regarding the`plotProfile` function (in the paper and in the function documentation as well).

Re: performance issues of `importPEBamFiles`, I am aware of it and I already thought about parallelizing in case of a single bam file. It has been on my TODO list before the next Bioconductor release. Since you mention it here, is it something I need to fix before approval for publication in JOSS ?

Thanks again for your review!
