### 2021-02-25 8ef988d
attempt to fix missing binaries for pkgdown build
