+++ .github/workflows/pkgdown.yaml
@@ -34,6 +34,7 @@ jobs:
 
       - name: Install dependencies
         run: |
+          options(pkgType="binary")
           remotes::install_deps(dependencies = TRUE)
           install.packages("pkgdown", type = "binary")
         shell: Rscript {0}
