### 2021-02-25 #3009 @fgeier
@js2264 @csoneson Thank you for the improved docs. Your submission meets the criteria for acceptance from my side. Concerning the performance, I am looking forward to the next releases. Thanks again for the nice package, I will integrate it in my work!
