+++ paper/paper.bib
@@ -31,6 +31,8 @@
 	title = {{Distinctive regulatory architectures of germline-active and somatic genes in C. elegans}},
 	journal = {Genome Research},
 	year = {2020},
+	volume = {30},
+	pages = {1752--1765},
 	month = {Oct},
 	issn = {1088-9051},
 	publisher = {Cold Spring Harbor Lab},
