### 2021-02-25 #3009 @csoneson
👋 @js2264 - I have sent a pull request with some minor fixes to the paper. Please have a look and merge. The next steps are as follows:
- [ ] Make a tagged release of the software, and list the version tag of the archived version here.
- [ ] Check that the archival deposit (e.g., in Zenodo) has the correct metadata - in particular, the title and author list should match those of the paper. You may also add the authors' ORCID.
- [ ] Please list the DOI of the Zenodo archive here.
