### 2021-02-25 7a2884e
bump to version 1.2.0
