+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: VplotR
 Type: Package
 Title: Set of tools to make V-plots and compute footprint profiles
-Version: 1.1.0
+Version: 1.2
 Date: 2020-05-03
 Encoding: UTF-8
 Authors@R: 
