### 2021-02-25 #3009 @js2264
Release is v1.2.0, I created a new tag for it (https://github.com/js2264/VplotR/tree/v1.2.0)
The package is now available in Zenodo (10.5281/zenodo.4562440). I checked the info and everything shoul dbe ok!
