### 2021-02-26 #3009 @kthyng
@fgeier We need all of the boxes checked to finish up reviews. This is our metric for the review and want them ticked off for the review to be finished.

For
> Performance: If there are any performance claims of the software, have they been confirmed? (If there are no claims, please check off this item.)
if there are no claims, you can check this off.

For
>  State of the field: Do the authors describe how this software compares to other commonly-used packages?
If this is not done in the paper, the author should fix this. If there is no comparable software, that should be stated and then the box can be checked off.
