### 2021-02-26 #3009 @kthyng
Looks like everything else is in order to finalize publication (version updated, Zenodo archive looks good, and paper is fine)!
