+++ paper/paper.bib
@@ -92,3 +92,15 @@
 	publisher = {The Author 2016Oxford University Press.},
 	doi = {10.1093/bioinformatics/btw024}
 }
+
+@incollection{Beati2020,
+	author = {Beati, Paula and Chereji, Răzvan V.},
+	title = {{Creating 2D Occupancy Plots Using plot2DO}},
+	booktitle = {{Stem Cell Transcriptional Networks}},
+	journal = {SpringerLink},
+	pages = {93--108},
+	year = {2020},
+	publisher = {Humana, New York, NY},
+	address = {New York, NY, USA},
+	doi = {10.1007/978-1-0716-0301-7_5}
+}
