### 2021-02-26 #3009 @js2264
@kthyng @fgeier About other commonly-used packages: I've just added a sentence about 2 tools that already exist. I have not changed the content of the code itself, only `paper.md` and `paper.bib`. Should I still re-create the Zenodo archive?
