### 2021-02-26 #3009 @kthyng
@js2264 no you don't need to update the archive.
